--- src/main/java/com/br/alura/controller/TopicosController.java
@@ -1,16 +1,0 @@
-package com.br.alura.controller;
-
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/topicos")
-public class TopicosController {
-
-    @PostMapping
-    public void cadastrar(){
-
-    }
-
-}
